+++ itextsharp.kernel/itextsharp/kernel/crypto/securityhandler/StandardHandlerUsingStandard40.cs
@@ -263,7 +263,7 @@ namespace iTextSharp.Kernel.Crypto.Securityhandler
 		private int GetKeyLength(PdfDictionary encryptionDict)
 		{
 			int? keyLength = encryptionDict.GetAsInt(PdfName.Length);
-			return keyLength != null ? keyLength : 40;
+			return keyLength != null ? (int)keyLength : 40;
 		}
 	}
 }
